ekoupdate

+++ resources/views/admin/feed/show_feed.blade.php
@@ -1,5 +1,4 @@
 @extends('admin.index')
-
 @section('data')
 <div class="container">
 
@@ -22,7 +21,8 @@
             @foreach ($feeds as $feed)
                 
             <tr>
-              <th scope="row">{{$feed->id}}</th>
+              {{-- <th scope="row">{{$feed->id}}</th> --}}
+              <th scope="row">{{$loop->iteration}}</th>
               <th><img style="width: 200px" src="{{asset('/images/feed/'.$feed->image_feed) }}" class="card-img-top fade-1" alt="..."></th>
               <td>{{$feed->link}}</td>
               <td>
@@ -40,7 +40,7 @@
               </td>
               <td>
                 <a href="/admin/update_feed/{{$feed->id}}" class="btn btn-warning">update</a>
-                <a href="/admin/delete_feed/{{$feed->id}}" class="btn btn-info">delete</a>
+                <a onclick="return confirm('Are you sure?')" href="/admin/delete_feed/{{$feed->id}}" class="btn btn-info">delete</a>
               </td>
             </tr>
             @endforeach
